refactor: move client to zoot
