+++ README.md
@@ -8,7 +8,7 @@ ZOOT Plus 前端！
 - 作业格式：[战斗流程协议](https://maa.plus/docs/zh-cn/protocol/copilot-schema.html)
 - i18n：[i18n/README.md](src/i18n/README.md)
 
-更新 zoot-plus-client 时，需要在 [Tags](https://github.com/ZOOT-Plus/zoot-plus-client-ts/tags) 中复制版本号，然后替换掉 `package.json` 中的 `maa-copilot-client` 版本号，再运行 `pnpm install` 安装依赖
+更新 zoot-plus-client 时，需要在 [Tags](https://github.com/ZOOT-Plus/zoot-plus-client-ts/tags) 中复制版本号，然后替换掉 `package.json` 中的 `zoot-plus-client` 版本号，再运行 `pnpm install` 安装依赖
 
 ## 开发流程
