+++ package.json
@@ -21,72 +21,72 @@
     "scripts:update-prof-icons": "pnpm install && tsx scripts/update-prof-icons.ts"
   },
   "dependencies": {
-    "@blueprintjs/core": "^6.16.0",
-    "@blueprintjs/icons": "^6.11.0",
-    "@blueprintjs/select": "^6.3.0",
+    "@blueprintjs/core": "^6.17.2",
+    "@blueprintjs/icons": "^6.13.0",
+    "@blueprintjs/select": "^6.3.3",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@iconify/icons-simple-icons": "^1.2.22",
+    "@iconify/icons-simple-icons": "^1.2.74",
     "@iconify/react": "^3.2.2",
-    "@sentry/react": "7.27.0",
-    "@sentry/tracing": "7.27.0",
-    "@sentry/utils": "7.27.0",
-    "ajv": "^8.11.0",
+    "@sentry/react": "7.120.4",
+    "@sentry/tracing": "7.120.4",
+    "@sentry/utils": "7.120.4",
+    "ajv": "^8.20.0",
     "ajv-i18n": "^4.2.0",
-    "camelcase-keys": "^8.0.1",
-    "clsx": "^1.2.0",
-    "dayjs": "^1.11.3",
+    "camelcase-keys": "^8.0.2",
+    "clsx": "^1.2.1",
+    "dayjs": "^1.11.21",
     "fuse.js": "^6.6.2",
-    "immer": "^10.1.1",
-    "jotai": "^2.7.0",
+    "immer": "^10.2.0",
+    "jotai": "^2.20.1",
     "jotai-devtools": "^0.11.0",
-    "jotai-immer": "^0.4.1",
+    "jotai-immer": "^0.4.3",
     "linkify-react": "^3.0.4",
     "linkifyjs": "^3.0.5",
-    "lodash-es": "^4.17.21",
-    "maa-copilot-client": "https://github.com/ZOOT-Plus/zoot-plus-client-ts.git#0.1.0-SNAPSHOT.999.sha.d286f0d",
+    "lodash-es": "^4.18.1",
     "mitt": "^3.0.1",
     "normalize.css": "^8.0.1",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-ga-neo": "^2.2.0",
-    "react-hook-form": "^7.33.1",
+    "react-hook-form": "^7.81.0",
     "react-markdown": "^10.1.0",
     "react-rating": "^2.0.5",
-    "react-resizable-panels": "^2.1.8",
-    "react-rnd": "^10.4.1",
-    "react-router-dom": "6",
-    "react-spring": "^9.4.5",
-    "react-use": "^17.4.0",
-    "remark-breaks": "^3.0.2",
+    "react-resizable-panels": "^2.1.9",
+    "react-rnd": "^10.5.3",
+    "react-router-dom": "^6.30.4",
+    "react-spring": "^9.7.5",
+    "react-use": "^17.6.1",
+    "remark-breaks": "^4.0.0",
     "remark-gfm": "^4.0.1",
-    "snakecase-keys": "^5.4.4",
-    "swr": "^2.2.5",
+    "snakecase-keys": "^5.5.0",
+    "swr": "^2.4.2",
     "tslib": "^2.8.1",
-    "type-fest": "^4.40.1",
+    "type-fest": "^4.41.0",
     "unfetch": "^4.2.0",
-    "zod": "^4.4.3"
+    "zod": "^4.4.3",
+    "zoot-plus-client": "github:ZOOT-Plus/zoot-plus-client-ts#0.1.0-featurevideocopilot.1002.sha.dc6d4dc"
   },
   "devDependencies": {
-    "@hookform/devtools": "^4.1.1",
-    "@types/lodash-es": "^4.17.6",
-    "@types/node": "^24.0.0",
-    "@types/pinyin": "^2.10.0",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@hookform/devtools": "^4.4.0",
+    "@types/lodash-es": "^4.17.12",
+    "@types/node": "^24.13.3",
+    "@types/pinyin": "^2.10.2",
+    "@types/react": "^18.3.31",
+    "@types/react-dom": "^18.3.7",
     "@vitejs/plugin-react": "^6.0.3",
-    "autoprefixer": "^10.4.7",
-    "less": "^4.1.2",
+    "autoprefixer": "^10.5.2",
+    "less": "^4.6.7",
     "oxfmt": "^0.56.0",
-    "oxlint": "^1.71.0",
+    "oxlint": "^1.73.0",
     "oxlint-tsgolint": "^0.23.0",
-    "pinyin": "^3.0.0-alpha.5",
-    "sharp": "^0.34.2",
+    "pinyin": "^3.1.0",
+    "sharp": "^0.34.5",
     "simplebig": "^0.0.3",
-    "tailwindcss": "^3.1.4",
-    "tsx": "^4.0.0",
-    "typescript": "7.0.1-rc",
-    "vite": "^8.1.0"
+    "tailwindcss": "^3.4.19",
+    "tsx": "^4.23.0",
+    "typescript": "7.0.2",
+    "vite": "^8.1.4"
   }
 }
