+++ pnpm-workspace.yaml
@@ -1,14 +1,9 @@
 allowBuilds:
   esbuild: true
-  maa-copilot-client: true
+  zoot-plus-client: true
   nodejieba: true
   sharp: true
   unrs-resolver: true
-
-# Rolldown(Vite 8）无法解析 @popperjs/core 经 export* 重导出的 placements，
-# 这里给 Blueprint v6 legacy Popover 依赖的 popper 打补丁绕过（详见 vite.config.ts TODO）。
-patchedDependencies:
-  '@popperjs/core@2.11.8': patches/@popperjs__core@2.11.8.patch
 
 overrides:
   react-rating>@types/react: ^18.0.0
