+++ src/models/operation-set.ts
@@ -1,4 +1,4 @@
-import { CopilotSetListRes, CopilotSetRes } from 'maa-copilot-client'
+import { CopilotSetListRes, CopilotSetRes } from 'zoot-plus-client'
 
 export type OperationSet = CopilotSetRes
 export type OperationSetListItem = CopilotSetListRes
