+++ src/styles/blueprint-icons.css
@@ -1,0 +1,14 @@
+/* ponytail: 本地裁剪版 @blueprintjs/icons 字体声明。
+   原版 blueprint-icons.css 的 @font-face 同时引用 ttf/eot/woff2/woff/svg 五种格式，
+   Vite 会把所有 url() 引用都作为 asset 产出（光 svg 就 ~970KB，整体 ~1.8MB）。
+   现代浏览器只需 woff2，这里仅保留 woff2，构建产物约 ~104KB。
+   升级 @blueprintjs/icons 时若其字体文件名/哈希变化，同步更新下方 url()。 */
+@font-face {
+  font-family: 'blueprint-icons-16';
+  src: url('../../node_modules/@blueprintjs/icons/lib/css/blueprint-icons-16.woff2') format('woff2');
+}
+
+@font-face {
+  font-family: 'blueprint-icons-20';
+  src: url('../../node_modules/@blueprintjs/icons/lib/css/blueprint-icons-20.woff2') format('woff2');
+}
