+++ vite.config.ts
@@ -9,14 +9,6 @@ import { generateTranslations } from './scripts/generate-translations'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
-// TODO(blueprint-v6 / @popperjs/core 补丁): 移除 patches/@popperjs__core@2.11.8.patch
-// 当前 Rolldown(Vite 8 打包器）在 app 打包下无法解析 @popperjs/core 经 `export *` 重导出、
-// 且带 `/*#__PURE__*/` 注解的 `placements`，导致 Blueprint v6 legacy Popover 的
-// `export { placements as PopperPlacements }` 构建失败（MISSING_EXPORT）。
-// 注意：rolldown#9122/#9958 只修了 preserveModules 库模式（已在 1.1.3 中，但不覆盖本场景）。
-// 待 Rolldown 发布覆盖 app-bundle export* 的修复后：删除 patches/ 里的补丁文件 +
-// `pnpm-workspace.yaml` 里的 patchedDependencies 条目，重新 `pnpm install && pnpm build` 验证即可。
-
 // https://vitejs.dev/config/
 export default defineConfig(({ command, mode }) => {
   // Load env file based on `mode` in the current working directory.
@@ -42,7 +34,7 @@ export default defineConfig(({ command, mode }) => {
         output: {
           codeSplitting: {
             groups: [
-              { name: 'maacopilot', test: /maa-copilot-client/ },
+              { name: 'zootplusclient', test: /zoot-plus-client/ },
               { name: 'react', test: /node_modules[\\/](react|react-dom|react-router-dom)[\\/]/ },
               {
                 name: 'reactplugins',
